requirements.txt
# Deep Audio Detection System - Requirements
# Portfolio-Ready with Version Constraints

# Core ML & Audio Processing
openai-whisper>=20230314
TTS>=0.22.0
librosa>=0.10.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
lightgbm>=4.0.0

# Web Framework
streamlit>=1.28.0
plotly>=5.17.0

# Audio I/O
soundfile>=0.12.0
pydub>=0.25.0

# Deep Learning
torch>=2.0.0
torchaudio>=2.0.0
